+++ src/Aspire.Hosting.Azure/AzureAppConfigurationResource.cs
@@ -3,9 +3,20 @@
 
 namespace Aspire.Hosting.ApplicationModel;
 
+/// <summary>
+/// Represents an Azure App Configuration resource.
+/// </summary>
+/// <param name="name">The name of the resource.</param>
 public class AzureAppConfigurationResource(string name) : Resource(name), IAzureResource, IResourceWithConnectionString
 {
+    /// <summary>
+    /// Gets or sets the endpoint for the Azure App Configuration resource.
+    /// </summary>
     public string? Endpoint { get; set; }
 
+    /// <summary>
+    /// Gets the connection string for the Azure App Configuration resource.
+    /// </summary>
+    /// <returns>The connection string for the Azure App Configuration resource.</returns>
     public string? GetConnectionString() => Endpoint;
 }
