+++ src/Aspire.Hosting.Azure/AzureBlobStorageResource.cs
@@ -1,0 +1,26 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Aspire.Hosting.ApplicationModel;
+
+/// <summary>
+/// Represents a resource that is stored in Azure Blob Storage.
+/// </summary>
+/// <param name="name">The name of the resource.</param>
+/// <param name="storage">The <see cref="AzureStorageResource"/> that the resource is stored in.</param>
+public class AzureBlobStorageResource(string name, AzureStorageResource storage) : Resource(name),
+    IAzureResource,
+    IResourceWithConnectionString,
+    IResourceWithParent<AzureStorageResource>
+{
+    /// <summary>
+    /// Gets the parent AzureStorageResource of this AzureBlobStorageResource.
+    /// </summary>
+    public AzureStorageResource Parent => storage;
+
+    /// <summary>
+    /// Gets the connection string for the Azure Blob Storage resource.
+    /// </summary>
+    /// <returns>The connection string for the Azure Blob Storage resource.</returns>
+    public string? GetConnectionString() => Parent.GetBlobConnectionString();
+}
