+++ src/Aspire.Hosting.Azure/AzureResourceExtensions.cs
@@ -7,6 +7,9 @@ using Aspire.Hosting.ApplicationModel;
 
 namespace Aspire.Hosting;
 
+/// <summary>
+/// Provides extension methods for adding Azure resources to the application model.
+/// </summary>
 public static class AzureResourceExtensions
 {
     /// <summary>
