+++ src/Components/Aspire.Azure.Data.Tables/AspireTablesExtensions.cs
@@ -14,6 +14,9 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 
 namespace Microsoft.Extensions.Hosting;
 
+/// <summary>
+/// Provides extension methods for registering <see cref="TableServiceClient"/> as a singleton in the services provided by the <see cref="IHostApplicationBuilder"/>.
+/// </summary>
 public static class AspireTablesExtensions
 {
     private const string DefaultConfigSectionName = "Aspire:Azure:Data:Tables";
